Move EventListYearView to module views.web
